+++ src/test/java/chess/domain/chessboard/ChessBoardTest.java
@@ -11,7 +11,7 @@ import chess.domain.chesspiece.ChessPiece;
 import chess.domain.chesspiece.Pawn;
 import chess.domain.chesspiece.Rook;
 import chess.domain.position.Position;
-import chess.factory.BoardFactory;
+import chess.domain.factory.BoardFactory;
 import chess.view.OutputView;
 
 public class ChessBoardTest {
